requirements.txt
Flask==2.1.1
gunicorn==20.1.0
python-whois==0.7.3